Add files via upload
The table for the Solar Calc. This is form of a database, with several different information to calculate with real datas. The first sdtep is to finnished that with scripts. On behalf of Gemmini you can formulate your questions over the data later what is interestet. Samples of questionares in german languges will be provided.